requirements.txt
aniso8601==6.0.0
attrs==19.1.0
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.9.2
firebase-admin==2.16.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
flask-restplus==0.12.1
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
MarkupSafe==1.1.1
msgpack==0.6.1
mysqlclient==1.4.6
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
PyNaCl==1.4.0
pyrsistent==0.14.11
python-dateutil==2.8.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
